Update rules.md
Fixed multiple spelling errors (I think it should say "firing guides").  Also, I am not sure what "firing guides" means. 

Also, because the range ruler is in three parts it isn't very clear when you read the instructions if the range is a single section of the ruler of the entire thing.